exit status used

+++ Roboshop/components/frontend.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-set -e
+#set -e
 
 # yum install nginx -y
 # systemctl enable nginx
@@ -25,8 +25,17 @@ set -e
 echo frontend setup
 
 yum install nginx -y
+if [ $? -ne 0 ]; then
+  echo "Nginx install failed"
+  exit
+fi
 
 curl -s -L -o /tmp/frontend.zip "https://github.com/roboshop-devops-project/frontend/archive/main.zi"
+if [ $? -ne 0 ]; then
+  echo "Nginx install failed"
+  exit
+fi
+
 cd /usr/share/nginx/html
 rm -rf *
 unzip /tmp/frontend.zip
